producer/requirements.txt
requests==2.32.3
kafka-python==2.3.1
python-dotenv==1.0.1